refac(zk): move halo2 related to `zk/plonk/halo2/`
