+++ tachyon/zk/lookup/test/BUILD.bazel
@@ -6,7 +6,7 @@ tachyon_cc_library(
     hdrs = ["compress_expression_test_setting.h"],
     visibility = ["//tachyon/zk/lookup:__pkg__"],
     deps = [
-        "//tachyon/zk/base/halo2:halo2_prover_test",
         "//tachyon/zk/expressions/evaluator:simple_evaluator",
+        "//tachyon/zk/plonk/halo2:halo2_prover_test",
     ],
 )
