+++ tachyon/zk/plonk/circuit/examples/circuit_test.h
@@ -4,7 +4,7 @@
 #include <utility>
 #include <vector>
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
