+++ tachyon/zk/plonk/halo2/constants.h
@@ -1,5 +1,5 @@
-#ifndef TACHYON_ZK_BASE_HALO2_CONSTANTS_H_
-#define TACHYON_ZK_BASE_HALO2_CONSTANTS_H_
+#ifndef TACHYON_ZK_PLONK_HALO2_CONSTANTS_H_
+#define TACHYON_ZK_PLONK_HALO2_CONSTANTS_H_
 
 namespace tachyon::zk::halo2 {
 
@@ -8,4 +8,4 @@ constexpr char kVerifyingKeyStr[] = "Halo2-Verify-Key";
 
 }  // namespace tachyon::zk::halo2
 
-#endif  // TACHYON_ZK_BASE_HALO2_CONSTANTS_H_
+#endif  // TACHYON_ZK_PLONK_HALO2_CONSTANTS_H_
