+++ tachyon/zk/plonk/halo2/pinned_verifying_key_unittest.cc
@@ -1,8 +1,8 @@
-#include "tachyon/zk/plonk/keys/halo2/pinned_verifying_key.h"
+#include "tachyon/zk/plonk/halo2/pinned_verifying_key.h"
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk::halo2 {
