+++ tachyon/zk/plonk/permutation/permutation_proving_key_unittest.cc
@@ -9,7 +9,7 @@
 #include "gtest/gtest.h"
 
 #include "tachyon/base/buffer/vector_buffer.h"
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
