+++ tachyon/zk/plonk/permutation/permutation_table_store_unittest.cc
@@ -5,7 +5,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
 namespace {
