+++ tachyon/zk/plonk/permutation/unpermuted_table_unittest.cc
@@ -10,7 +10,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
