+++ tachyon/zk/plonk/prover/argument_unittest.cc
@@ -11,7 +11,7 @@
 #include "gtest/gtest.h"
 
 #include "tachyon/base/containers/container_util.h"
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
 namespace {
