+++ tachyon/zk/plonk/prover/synthesizer_unittest.cc
@@ -10,9 +10,9 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
 #include "tachyon/zk/plonk/circuit/examples/simple_circuit.h"
-#include "tachyon/zk/plonk/keys/halo2/pinned_verifying_key.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/halo2/pinned_verifying_key.h"
 
 namespace tachyon::zk {
 namespace {
