+++ tachyon/zk/plonk/prover/witness_collection_unittest.cc
@@ -8,9 +8,9 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
 #include "tachyon/zk/base/value.h"
 #include "tachyon/zk/plonk/circuit/examples/simple_circuit.h"
+#include "tachyon/zk/plonk/halo2/halo2_prover_test.h"
 
 namespace tachyon::zk {
 namespace {
